ID type change
changed id type from Integer to Long

+++ src/main/java/com/ISS/Booking_iss_tim21/controller/NotificationController.java
@@ -34,7 +34,7 @@ public class NotificationController {
     }
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<Void> deleteCourse(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteCourse(@PathVariable Long id) {
 
         Notification course = service.findOne(id);
